add new `QueryParams` type for better semantics
